@@ -117,7 +117,7 @@ public:
 					name = value;
 
 				}else if(attr == "namespace"){
-					//go ontology, BP,MF,CC
+					//go ontology, BIOLOGICAL_PROCESS,MOLECULAR_FUNCTION,CELLULAR_COMPONENT
 					ontology = value;
 
 				}else if(attr == "def"){
