@@ -8,6 +8,11 @@ file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 #define ANNOTATION_PARSER_FACTORY
 
 #include <AnnotationParserInterface.hpp>
+#include <DisallowedSetEvidencePolicy.hpp>
+#include <EntrezGene2GoAnnotationParser.hpp>
+#include <GoaAnnotationParser.hpp>
+#include <GafAnnotationParser.hpp>
+#include <MgiAnnotationParser.hpp>
 #include <vector>
 #include <string>
 
@@ -20,79 +25,37 @@ file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 	  parsers to be easily added to a larger system and switched at runtime.
 
 */
+enum class AnnotationParserType { ENTREZ_ANNO_PARSER, GOA_ANNO_PARSER, GAF_ANNO_PARSER, MGI_ANNO_PARSER };
 class AnnotationParserFactory{
 
-private:
-	//! A list of strings which correspond to parsers
-	/*!
-		This list holds the names of parsers added to the factory. These are the
-		  keys used to query a parser.
-	*/
-	std::vector<std::string> _names;
-
-	//! A list of parsers which are currently held in the factory.
-	/*!
-		This list holds the instances of AnnotationParserInterface that can be returned.
-	*/
-	std::vector<AnnotationParserInterface*> _parsers;
 
 public:
 
-	//! Class constructor
-	/*!
-		This constructor initializes the private lists to empty vectors.
-	*/
-	inline AnnotationParserFactory(){
-		//initialize vectors
-		_names   = std::vector<std::string>();
-		_parsers = std::vector<AnnotationParserInterface*>();
-	}//end Constructor
-
-	//! Class destructor
-	/*!
-		This destructor clears the names vector. It also deletes each parser pointer explicitly.
-		  Finally it clears the parser list.
-	*/
-	inline ~AnnotationParserFactory(){
-		//clear vectors
-		_names.clear();
-
-		//explicitly delete internal pointers
-		for(std::size_t i = 0; i < _parsers.size();++i){
-			delete _parsers.at(i);
-		}
-		_parsers.clear();
-
-	}//end destructor
+  // Cannot create object.
+	AnnotationParserFactory() = delete;
+	~AnnotationParserFactory() = delete;
 
 
-	//! A Method to add a parser to the factory.
-	/*!
-		This method adds a pointer to a parser and a string to the factory.
-		  This string is used to query the appropriate parser.
-	*/
-	inline void addParser(std::string name,AnnotationParserInterface* parser){
-		_names.push_back(name);
-		_parsers.push_back(parser);
-	}//end method addParser
+  [[nodiscard]] static std::unique_ptr<AnnotationParserInterface> createAnnotationParser( AnnotationParserType parser_type,
+                                                                                          const EvidencePolicyInterface& policy = DisallowedSetEvidencePolicy()) {
 
+    switch(parser_type) {
 
-	//! A method to return a parser based on a query string.
-	/*!
-		If the string supplied matches one of the keys in the database, the
-		 appropriate parser will be returned. If not, a NULL pointer is returned.
-		 The calling environment must check against NULL before using.
-	*/
-	inline AnnotationParserInterface* getParser(std::string name){
-		//check for the string
-		for(std::size_t i = 0; i < _names.size(); ++i){
-			if(_names.at(i).compare(name) == 0){
-				return _parsers.at(i)->clone();
-			}
-		}
-		//if not found return null
-		return NULL;
-	}//end method getParser
+      case AnnotationParserType::ENTREZ_ANNO_PARSER:
+        return std::make_unique<EntrezGene2GoAnnotationParser>(policy);
+
+      case AnnotationParserType::GOA_ANNO_PARSER:
+        return std::make_unique<GoaAnnotationParser>(policy);
+
+      case AnnotationParserType::GAF_ANNO_PARSER:
+        return std::make_unique<GafAnnotationParser>(policy);
+
+      case AnnotationParserType::MGI_ANNO_PARSER:
+        return std::make_unique<MgiAnnotationParser>(policy);
+
+    }
+
+  }
 
 };
 #endif
