@@ -20,20 +20,25 @@ Implements AnnotationParserInterface.
 For now, the important aspects of the GAF file and GOA file are the same.
 The GafAnnotationParser inherits all functionality from GoaAnnotationParser. This may change in the future.
 */
-class GafAnnotationParser : public GoaAnnotationParser{
+class GafAnnotationParser : public GoaAnnotationParser {
 
 public:
 	//! A default constructor method for creating the parser
 	/*!
 		Creates the parser
 	*/
-	inline GafAnnotationParser() : GoaAnnotationParser(){}
+	GafAnnotationParser() : GoaAnnotationParser() {}
 
 	//! A parameterized constructor method for creating the parser with a policy.
 	/*!
 		Creates the parser with a custom policy
 	*/
-	inline GafAnnotationParser(EvidencePolicyInterface* policy) : GoaAnnotationParser(policy){}
+	explicit GafAnnotationParser(const EvidencePolicyInterface& policy) : GoaAnnotationParser(policy) {}
+
+  ~GafAnnotationParser() override = default;
+
+  [[nodiscard]] std::unique_ptr<AnnotationParserInterface> clone() const override { return std::make_unique<GafAnnotationParser>(getPolicy()); }
+
 
 };
 
