@@ -9,14 +9,12 @@ file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <string>
 
-//! Number of ontologies, 3 + 1 error code
-#define NUM_ONTOLOGIES 4
 
 //! Number of evidence codes, 22 + 1 error code
-#define NUM_EVIDENCES 23
+static const constexpr size_t NUM_EVIDENCES = 23;
 
 //! Number of relationships codes, 5 + 1 error code
-#define NUM_RELATIONSHIPS 6
+static const constexpr size_t NUM_RELATIONSHIPS = 6;
 
 
 //! GO namespaces
@@ -26,28 +24,25 @@ file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 namespace GO{
 
 	//! function that returns strings representing the root ontology term biological_process
-	inline std::string getRootTermBP(){
-		return "GO:0008150";
-	}
+  static const constexpr char* ROOT_TERM_BIOLOICAL_PROCESS = "GO:0008150";
+	[[nodiscard]] std::string getRootTermBP(){ return ROOT_TERM_BIOLOICAL_PROCESS; }
 
 	//! function that returns strings representing the root ontology term molecular_function
-	inline std::string getRootTermMF(){
-		return "GO:0003674";
-	}
+  static const constexpr char* ROOT_TERM_MOLECULAR_FUNCTION = "GO:0003674";
+	[[nodiscard]] std::string getRootTermMF(){ return ROOT_TERM_MOLECULAR_FUNCTION; }
 	
 	//! function that returns strings representing the root ontology term cellular_component
-	inline std::string getRootTermCC(){
-		return "GO:0005575";
-	}
+  static const constexpr char* ROOT_TERM_CELLULAR_COMPONENT = "GO:0005575";
+	[[nodiscard]] std::string getRootTermCC(){ return ROOT_TERM_CELLULAR_COMPONENT; }
 
 	//! Ontology enum type
 	/*!
 		This enum defines a type for sub-ontologies.
 	*/
-	enum Onto{
-		BP=0,
-		MF=1,
-		CC=2,
+	enum class Ontology {
+		BIOLOGICAL_PROCESS=0,
+		MOLECULAR_FUNCTION=1,
+		CELLULAR_COMPONENT=2,
 		ONTO_ERROR=3
 	};
 
@@ -55,39 +50,61 @@ namespace GO{
 	/*!
 		These are string representations of the sub-ontologies. Take from obo go files.
 	*/
-	static std::string ontologyStringCodes[NUM_ONTOLOGIES] = {
-		"biological_process",
-		"molecular_function",
-		"cellular_component",
-		"ONTOLOGY_ERROR"
-	};
+
+	static const constexpr char* ONTOLOGY_BIOLOGICAL_PROCESS_TEXT = "biological_process";
+  static const constexpr char* ONTOLOGY_MOLECULAR_FUNCTION_TEXT = "molecular_function";
+  static const constexpr char* ONTOLOGY_CELLULAR_COMPONENT_TEXT = "cellular_component";
+  static const constexpr char* ONTOLOGY_ERROR_TEXT = "ONTOLOGY_ERROR";
+
+  struct OntologyText { const char* text; Ontology onto; };
+	static const constexpr OntologyText ontology_text[] = {
+
+	    { ONTOLOGY_BIOLOGICAL_PROCESS_TEXT, Ontology::BIOLOGICAL_PROCESS},
+      { ONTOLOGY_MOLECULAR_FUNCTION_TEXT, Ontology::MOLECULAR_FUNCTION},
+      { ONTOLOGY_CELLULAR_COMPONENT_TEXT, Ontology::CELLULAR_COMPONENT},
+      { ONTOLOGY_ERROR_TEXT, Ontology::ONTO_ERROR}
+
+  };
+
 
 	//! A method for returning the ontology code based on string
 	/*!
 		This method takes a string and returns the proper enum
 	*/
-	inline Onto ontologyStringToCode(std::string code){
-		std::size_t i = 0;
-		for(; i < NUM_ONTOLOGIES - 1; ++i){
+	[[nodiscard]] Ontology ontologyStringToCode(const std::string& code) {
+
+		for(auto const& [text, ontology] : ontology_text) {
 			//return the index if matching, cast as enum
-			if(code.compare(ontologyStringCodes[i]) == 0){
-				return static_cast<Onto>(i);
+			if(code == text) {
+
+				return ontology;
+
 			}
+
 		}
 		//return error enum if not found
-		return static_cast<Onto>(i);
+		return Ontology::ONTO_ERROR;
+
 	}
 
 	//! A method for returning a human readable string from the ontology code
 	/*!
 		This method takes an ontology enum value and returns a string
 	*/
-	inline std::string ontologyToString(const Onto &o){
-		if (o < GO::BP || o >= GO::ONTO_ERROR){
-			return ontologyStringCodes[GO::ONTO_ERROR];
-		}else{
-			return ontologyStringCodes[o];
-		}
+[[nodiscard]] std::string ontologyToString(Ontology onto) {
+
+    for(auto const& [text, ontology] : ontology_text) {
+
+      if(ontology == onto) {
+
+        return text;
+
+      }
+
+    }
+
+    return ONTOLOGY_ERROR_TEXT;
+
 	}
 
 	//! Evcidence Code enum type
@@ -95,7 +112,7 @@ namespace GO{
 		This enum defines a type for evidence codes.
 		 Defined at http://www.geneontology.org/GO.evidence.shtml
 	*/
-	enum EvidenceCode{
+	enum class EvidenceCode{
 		//experimental
 		EXP=0,
 		IDA=1,
@@ -132,81 +149,100 @@ namespace GO{
 		ECODE_ERROR=22
 	};
 
+  enum class EvidenceType { EXPERIMENTAL, COMPUTATIONAL, AUTHOR, CURATOR, AUTO_ASSIGNED, OBSOLETE, ERROR };
 
 	//! Relationship code enum strings
 	/*!
 		These are string representations of the evidence codes for an annotation.
 	*/
-	static std::string evidenceStringCodes[NUM_EVIDENCES] = {
+  struct EvidenceText { const char* text; EvidenceCode code; EvidenceType type; };
+  static const constexpr char* EVIDENCE_ERROR_TEXT = "EVIDENCE_CODE_ERROR";
+	static const constexpr EvidenceText evidence_text[] = {
 
 		//experimental
-		"EXP",
-		"IDA",
-		"IPI",
-		"IMP",
-		"IGI",
-		"IEP",
+      { "EXP", EvidenceCode::EXP, EvidenceType::EXPERIMENTAL },
+      { "IDA", EvidenceCode::IDA, EvidenceType::EXPERIMENTAL },
+      { "IPI", EvidenceCode::IPI, EvidenceType::EXPERIMENTAL },
+      { "IMP", EvidenceCode::IMP, EvidenceType::EXPERIMENTAL },
+      { "IGI", EvidenceCode::IGI, EvidenceType::EXPERIMENTAL },
+      { "IEP", EvidenceCode::IEP, EvidenceType::EXPERIMENTAL },
 
 		//computationally assisted
-		"ISS",
-		"ISO",
-		"ISA",
-		"ISM",
-		"IGC",
-		"IBA",
-		"IBD",
-		"IKR",
-		"IRD",
-		"RCA",
+      { "ISS", EvidenceCode::ISS, EvidenceType::COMPUTATIONAL },
+      { "ISO", EvidenceCode::ISO, EvidenceType::COMPUTATIONAL },
+      { "ISA", EvidenceCode::ISA, EvidenceType::COMPUTATIONAL },
+      { "ISM", EvidenceCode::ISM, EvidenceType::COMPUTATIONAL },
+      { "IGC", EvidenceCode::IGC, EvidenceType::COMPUTATIONAL },
+      { "IBA", EvidenceCode::IBA, EvidenceType::COMPUTATIONAL },
+      { "IBD", EvidenceCode::IBD, EvidenceType::COMPUTATIONAL },
+      { "IKR", EvidenceCode::IKR, EvidenceType::COMPUTATIONAL },
+      { "IRD", EvidenceCode::IRD, EvidenceType::COMPUTATIONAL },
+      { "RCA", EvidenceCode::RCA, EvidenceType::COMPUTATIONAL },
 
 		//author statement
-		"TAS",
-		"NAS",
+      { "TAS", EvidenceCode::TAS, EvidenceType::AUTHOR },
+      { "NAS", EvidenceCode::NAS, EvidenceType::AUTHOR },
 
 		//Curator statement
-		"IC",
-		"ND",
+      { "IC", EvidenceCode::IC, EvidenceType::CURATOR },
+      { "ND", EvidenceCode::ND, EvidenceType::CURATOR },
 
 		//automatically assigned
-		"IEA",
+      { "IEA", EvidenceCode::IEA, EvidenceType::AUTO_ASSIGNED },
 
 		//obsolete evidence code
-		"NR",
+      { "NR", EvidenceCode::NR, EvidenceType::OBSOLETE },
 
 		//Error code
-		"EVIDENCE_CODE_ERROR"
+      { EVIDENCE_ERROR_TEXT, EvidenceCode::ECODE_ERROR, EvidenceType::ERROR }
 	};
 
 	//! A method for converting evidence code strings to enums
 	/*!
 		This method takes a string representing the evidence code and converts it to an enum.
 	*/
-	inline EvidenceCode evidenceStringToCode(std::string code){
-		std::size_t i = 0;
-		for(; i < NUM_EVIDENCES - 1; ++i){
+	[[nodiscard]] EvidenceCode evidenceStringToCode(const std::string& text_code){
+
+		for(auto const& [text, code, type] : evidence_text) {
 			//return the index if matching, cast as enum
-			if(code.compare(evidenceStringCodes[i]) == 0){
-				return static_cast<EvidenceCode>(i);
+			if(text_code == text) {
+
+				return code;
+
 			}
+
 		}
 		//return evidence error if not found
-		return static_cast<EvidenceCode>(i);
+		return EvidenceCode::ECODE_ERROR;
+
 	}
 
 	//! A method for returning a human readable string from an evidence code
 	/*!
 		This method takes an evidence code enum value and returns a string
 	*/
-	inline std::string evidenceToString(const EvidenceCode &evidence){
-		return evidenceStringCodes[evidence];
-	}
+	[[nodiscard]] std::string evidenceToString(EvidenceCode evidence){
+
+    for(auto const& [text, code, type] : evidence_text) {
+      //return the index if matching, cast as enum
+      if (evidence == code) {
+
+        return text;
+
+      }
+
+    }
+    //return evidence error if not found
+    return EVIDENCE_ERROR_TEXT;
+
+  }
 
 
 	//! Relationship codes enum
 	/*!
 		This enum represents the relationship codes for ontology edges.
 	*/
-	enum Relationship{
+	enum class Relationship {
 		IS_A=0,
 		PART_OF=1,
 		REGULATES=2,
@@ -216,41 +252,63 @@ namespace GO{
 	};
 
 
-	//! Relationship code enum strings
-	/*!
-		These strings represent the enum codes for each relationship.
-	*/
-	static std::string relationshipStringCodes[NUM_RELATIONSHIPS] = {
-		"is_a",
-		"part_of",
-		"regulates",
-		"positively_regulates",
-		"negatively_regulates",
-		"RELATIONSHIP_ERROR"
+//! Relationship code enum strings
+/*!
+  These strings represent the enum codes for each relationship.
+*/
+
+struct RelationshipText { const char* text; Relationship type; };
+static const constexpr char* RELATIONSHIP_ERROR_TEXT = "RELATIONSHIP_ERROR";
+static const constexpr RelationshipText relationship_text[] = {
+
+    { "is_a", Relationship::IS_A },
+    { "part_of", Relationship::PART_OF },
+    { "regulates", Relationship::REGULATES },
+    { "positively_regulates", Relationship::POSITIVELY_REGULATES },
+    { "negatively_regulates", Relationship::NEGATIVELY_REGULATES },
+    { RELATIONSHIP_ERROR_TEXT, Relationship::REL_ERROR }
+
 	};
 
 	//! A method to convert relationship codes from string to enum
 	/*!
 		This method converts the string representation of a relationship to an enum.
 	*/
-	inline Relationship relationshipStringToCode(std::string code){
-		std::size_t i = 0;
-		for(; i < NUM_RELATIONSHIPS - 1; ++i){
+	[[nodiscard]] Relationship relationshipStringToCode(const std::string& code){
+
+		for(auto const& [text, relation] : relationship_text) {
 			//return the index if matching, cast as enum
-			if(code.compare(relationshipStringCodes[i]) == 0){
-				return static_cast<Relationship>(i);
+			if(code == text){
+
+				return relation;
+
 			}
+
 		}
 		//return error code if not found
-		return static_cast<Relationship>(i);
+		return Relationship::REL_ERROR;
+
 	}
 
 	//! A method for returning a human readable string from a Relationship
 	/*!
 		This method takes an evidence code enum pointer value and returns a string
 	*/
-	inline std::string relationshipToString(const Relationship &relationship){
-		return relationshipStringCodes[relationship];
+	[[nodiscard]] std::string relationshipToString(Relationship relationship){
+
+    for(auto const& [text, relation] : relationship_text) {
+      //return the index if matching, cast as enum
+      if(relation == relationship){
+
+        return text;
+
+      }
+
+    }
+    //return error code if not found
+		return RELATIONSHIP_ERROR_TEXT;
+
 	}
+
 }
 #endif
